requirements.txt
biopython==1.79
numpy==1.22.3
tqdm==4.64.0